+++ lib/src/avd_parser.dart
@@ -1,35 +1,41 @@
 // ignore_for_file: public_member_api_docs
 import 'dart:ui';
 
-import 'package:xml/xml.dart';
 import 'package:path_drawing/path_drawing.dart';
 import 'package:vector_math/vector_math_64.dart';
+import 'package:xml/xml.dart';
 
 import 'avd/xml_parsers.dart';
 import 'vector_drawable.dart';
 
 class DrawableAvdRoot extends DrawableRoot {
-  const DrawableAvdRoot(DrawableViewport viewBox, List<Drawable> children,
-      DrawableDefinitionServer definitions, DrawableStyle style)
-      : super(viewBox, children, definitions, style);
+  const DrawableAvdRoot(
+      String id,
+      DrawableViewport viewBox,
+      List<Drawable> children,
+      DrawableDefinitionServer definitions,
+      DrawableStyle style)
+      : super(id, viewBox, children, definitions, style);
 }
 
 /// An SVG Shape element that will be drawn to the canvas.
 class DrawableAvdPath extends DrawableShape {
-  const DrawableAvdPath(Path path, DrawableStyle style) : super(path, style);
+  const DrawableAvdPath(String? id, Path path, DrawableStyle style)
+      : super(id, path, style);
 
   /// Creates a [DrawableAvdPath] from an XML <path> element
   factory DrawableAvdPath.fromXml(XmlElement el) {
     final String d =
-        getAttribute(el.attributes, 'pathData', def: '', namespace: androidNS);
+        getAttribute(el.attributes, 'pathData', def: '', namespace: androidNS)!;
     final Path path = parseSvgPathData(d);
-    assert(path != null);
+    assert(path != null); // ignore: unnecessary_null_comparison
 
     path.fillType = parsePathFillType(el.attributes);
-    final DrawablePaint stroke = parseStroke(el.attributes, path.getBounds());
-    final DrawablePaint fill = parseFill(el.attributes, path.getBounds());
+    final DrawablePaint? stroke = parseStroke(el.attributes, path.getBounds());
+    final DrawablePaint? fill = parseFill(el.attributes, path.getBounds());
 
     return DrawableAvdPath(
+      getAttribute(el.attributes, 'id', def: ''),
       path,
       DrawableStyle(stroke: stroke, fill: fill),
     );
@@ -47,7 +53,7 @@ Drawable parseAvdElement(XmlElement el, Rect bounds) {
   }
   // TODO(dnfield): clipPath
   print('Unhandled element ${el.name.local}');
-  return const DrawableGroup(null, null);
+  return const DrawableGroup('', null, null);
 }
 
 /// Parses an AVD <group> element.
@@ -56,24 +62,23 @@ Drawable parseAvdGroup(XmlElement el, Rect bounds) {
   for (XmlNode child in el.children) {
     if (child is XmlElement) {
       final Drawable el = parseAvdElement(child, bounds);
-      if (el != null) {
-        children.add(el);
-      }
+      children.add(el);
     }
   }
 
   final Matrix4 transform = parseTransform(el.attributes);
 
-  final DrawablePaint fill = parseFill(el.attributes, bounds);
-  final DrawablePaint stroke = parseStroke(el.attributes, bounds);
+  final DrawablePaint? fill = parseFill(el.attributes, bounds);
+  final DrawablePaint? stroke = parseStroke(el.attributes, bounds);
 
   return DrawableGroup(
+    getAttribute(el.attributes, 'id', def: ''),
     children,
     DrawableStyle(
       stroke: stroke,
       fill: fill,
       groupOpacity: 1.0,
     ),
-    transform: transform?.storage,
+    transform: transform.storage,
   );
 }