+++ lib/svg.dart
@@ -5,6 +5,7 @@ import 'dart:typed_data';
 import 'dart:ui' show Picture;
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart' show AssetBundle;
 import 'package:flutter/widgets.dart';
 
@@ -12,6 +13,7 @@ import 'parser.dart';
 import 'src/picture_provider.dart';
 import 'src/picture_stream.dart';
 import 'src/render_picture.dart';
+import 'src/unbounded_color_filtered.dart';
 import 'src/vector_drawable.dart';
 
 /// Instance for [Svg]'s utility methods, which can produce a [DrawableRoot]
@@ -26,6 +28,12 @@ final Svg svg = Svg._();
 class Svg {
   Svg._();
 
+  /// A global override flag for [SvgPicture.cacheColorFilter].
+  ///
+  /// If this is null, the value in [SvgPicture.cacheColorFilter] is used. If it
+  /// is not null, it will override that value.
+  bool? cacheColorFilterOverride;
+
   /// Produces a [PictureInfo] from a [Uint8List] of SVG byte data (assumes UTF8 encoding).
   ///
   /// The `allowDrawingOutsideOfViewBox` parameter should be used with caution -
@@ -39,7 +47,7 @@ class Svg {
   Future<PictureInfo> svgPictureDecoder(
     Uint8List raw,
     bool allowDrawingOutsideOfViewBox,
-    ColorFilter colorFilter,
+    ColorFilter? colorFilter,
     String key,
   ) async {
     final DrawableRoot svgRoot = await fromSvgBytes(raw, key);
@@ -67,7 +75,7 @@ class Svg {
   Future<PictureInfo> svgPictureStringDecoder(
       String raw,
       bool allowDrawingOutsideOfViewBox,
-      ColorFilter colorFilter,
+      ColorFilter? colorFilter,
       String key) async {
     final DrawableRoot svg = await fromSvgString(raw, key);
     return PictureInfo(
@@ -128,12 +136,12 @@ class Svg {
 ///  * [PictureCache], which holds images that may be reused.
 Future<void> precachePicture(
   PictureProvider provider,
-  BuildContext context, {
-  Rect viewBox,
-  ColorFilter colorFilterOverride,
-  Color color,
-  BlendMode colorBlendMode,
-  PictureErrorListener onError,
+  BuildContext? context, {
+  Rect? viewBox,
+  ColorFilter? colorFilterOverride,
+  Color? color,
+  BlendMode? colorBlendMode,
+  PictureErrorListener? onError,
 }) {
   final PictureConfiguration config = createLocalPictureConfiguration(
     context,
@@ -143,14 +151,14 @@ Future<void> precachePicture(
     colorBlendMode: colorBlendMode,
   );
   final Completer<void> completer = Completer<void>();
-  PictureStream stream;
+  PictureStream? stream;
 
-  void listener(PictureInfo picture, bool synchronous) {
+  void listener(PictureInfo? picture, bool synchronous) {
     completer.complete();
     stream?.removeListener(listener);
   }
 
-  void errorListener(dynamic exception, StackTrace stackTrace) {
+  void errorListener(Object exception, StackTrace stackTrace) {
     if (onError != null) {
       onError(exception, stackTrace);
     } else {
@@ -201,7 +209,7 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   const SvgPicture(
     this.pictureProvider, {
-    Key key,
+    Key? key,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
@@ -211,6 +219,9 @@ class SvgPicture extends StatefulWidget {
     this.placeholderBuilder,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.colorFilter,
+    this.cacheColorFilter = false,
   }) : super(key: key);
 
   /// Instantiates a widget that renders an SVG picture from an [AssetBundle].
@@ -292,28 +303,34 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   SvgPicture.asset(
     String assetName, {
-    Key key,
+    Key? key,
     this.matchTextDirection = false,
-    AssetBundle bundle,
-    String package,
+    AssetBundle? bundle,
+    String? package,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
     this.alignment = Alignment.center,
     this.allowDrawingOutsideViewBox = false,
     this.placeholderBuilder,
-    Color color,
+    Color? color,
     BlendMode colorBlendMode = BlendMode.srcIn,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.cacheColorFilter = false,
   })  : pictureProvider = ExactAssetPicture(
-            allowDrawingOutsideViewBox == true
-                ? svgStringDecoderOutsideViewBox
-                : svgStringDecoder,
-            assetName,
-            bundle: bundle,
-            package: package,
-            colorFilter: _getColorFilter(color, colorBlendMode)),
+          allowDrawingOutsideViewBox == true
+              ? svgStringDecoderOutsideViewBox
+              : svgStringDecoder,
+          assetName,
+          bundle: bundle,
+          package: package,
+          colorFilter: svg.cacheColorFilterOverride ?? cacheColorFilter
+              ? _getColorFilter(color, colorBlendMode)
+              : null,
+        ),
+        colorFilter = _getColorFilter(color, colorBlendMode),
         super(key: key);
 
   /// Creates a widget that displays a [PictureStream] obtained from the network.
@@ -347,8 +364,8 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   SvgPicture.network(
     String url, {
-    Key key,
-    Map<String, String> headers,
+    Key? key,
+    Map<String, String>? headers,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
@@ -356,17 +373,23 @@ class SvgPicture extends StatefulWidget {
     this.matchTextDirection = false,
     this.allowDrawingOutsideViewBox = false,
     this.placeholderBuilder,
-    Color color,
+    Color? color,
     BlendMode colorBlendMode = BlendMode.srcIn,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.cacheColorFilter = false,
   })  : pictureProvider = NetworkPicture(
-            allowDrawingOutsideViewBox == true
-                ? svgByteDecoderOutsideViewBox
-                : svgByteDecoder,
-            url,
-            headers: headers,
-            colorFilter: _getColorFilter(color, colorBlendMode)),
+          allowDrawingOutsideViewBox == true
+              ? svgByteDecoderOutsideViewBox
+              : svgByteDecoder,
+          url,
+          headers: headers,
+          colorFilter: svg.cacheColorFilterOverride ?? cacheColorFilter
+              ? _getColorFilter(color, colorBlendMode)
+              : null,
+        ),
+        colorFilter = _getColorFilter(color, colorBlendMode),
         super(key: key);
 
   /// Creates a widget that displays a [PictureStream] obtained from a [File].
@@ -398,7 +421,7 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   SvgPicture.file(
     File file, {
-    Key key,
+    Key? key,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
@@ -406,16 +429,22 @@ class SvgPicture extends StatefulWidget {
     this.matchTextDirection = false,
     this.allowDrawingOutsideViewBox = false,
     this.placeholderBuilder,
-    Color color,
+    Color? color,
     BlendMode colorBlendMode = BlendMode.srcIn,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.cacheColorFilter = false,
   })  : pictureProvider = FilePicture(
-            allowDrawingOutsideViewBox == true
-                ? svgByteDecoderOutsideViewBox
-                : svgByteDecoder,
-            file,
-            colorFilter: _getColorFilter(color, colorBlendMode)),
+          allowDrawingOutsideViewBox == true
+              ? svgByteDecoderOutsideViewBox
+              : svgByteDecoder,
+          file,
+          colorFilter: svg.cacheColorFilterOverride ?? cacheColorFilter
+              ? _getColorFilter(color, colorBlendMode)
+              : null,
+        ),
+        colorFilter = _getColorFilter(color, colorBlendMode),
         super(key: key);
 
   /// Creates a widget that displays a [PictureStream] obtained from a [Uint8List].
@@ -444,7 +473,7 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   SvgPicture.memory(
     Uint8List bytes, {
-    Key key,
+    Key? key,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
@@ -452,16 +481,22 @@ class SvgPicture extends StatefulWidget {
     this.matchTextDirection = false,
     this.allowDrawingOutsideViewBox = false,
     this.placeholderBuilder,
-    Color color,
+    Color? color,
     BlendMode colorBlendMode = BlendMode.srcIn,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.cacheColorFilter = false,
   })  : pictureProvider = MemoryPicture(
-            allowDrawingOutsideViewBox == true
-                ? svgByteDecoderOutsideViewBox
-                : svgByteDecoder,
-            bytes,
-            colorFilter: _getColorFilter(color, colorBlendMode)),
+          allowDrawingOutsideViewBox == true
+              ? svgByteDecoderOutsideViewBox
+              : svgByteDecoder,
+          bytes,
+          colorFilter: svg.cacheColorFilterOverride ?? cacheColorFilter
+              ? _getColorFilter(color, colorBlendMode)
+              : null,
+        ),
+        colorFilter = _getColorFilter(color, colorBlendMode),
         super(key: key);
 
   /// Creates a widget that displays a [PictureStream] obtained from a [String].
@@ -490,7 +525,7 @@ class SvgPicture extends StatefulWidget {
   /// If [excludeFromSemantics] is true, then [semanticLabel] will be ignored.
   SvgPicture.string(
     String bytes, {
-    Key key,
+    Key? key,
     this.width,
     this.height,
     this.fit = BoxFit.contain,
@@ -498,16 +533,22 @@ class SvgPicture extends StatefulWidget {
     this.matchTextDirection = false,
     this.allowDrawingOutsideViewBox = false,
     this.placeholderBuilder,
-    Color color,
+    Color? color,
     BlendMode colorBlendMode = BlendMode.srcIn,
     this.semanticsLabel,
     this.excludeFromSemantics = false,
+    this.clipBehavior = Clip.hardEdge,
+    this.cacheColorFilter = false,
   })  : pictureProvider = StringPicture(
-            allowDrawingOutsideViewBox == true
-                ? svgStringDecoderOutsideViewBox
-                : svgStringDecoder,
-            bytes,
-            colorFilter: _getColorFilter(color, colorBlendMode)),
+          allowDrawingOutsideViewBox == true
+              ? svgStringDecoderOutsideViewBox
+              : svgStringDecoder,
+          bytes,
+          colorFilter: svg.cacheColorFilterOverride ?? cacheColorFilter
+              ? _getColorFilter(color, colorBlendMode)
+              : null,
+        ),
+        colorFilter = _getColorFilter(color, colorBlendMode),
         super(key: key);
 
   /// The default placeholder for a SVG that may take time to parse or
@@ -515,38 +556,36 @@ class SvgPicture extends StatefulWidget {
   static WidgetBuilder defaultPlaceholderBuilder =
       (BuildContext ctx) => const LimitedBox();
 
-  static ColorFilter _getColorFilter(Color color, BlendMode colorBlendMode) =>
-      color == null
-          ? null
-          : ColorFilter.mode(color, colorBlendMode ?? BlendMode.srcIn);
+  static ColorFilter? _getColorFilter(Color? color, BlendMode colorBlendMode) =>
+      color == null ? null : ColorFilter.mode(color, colorBlendMode);
 
   /// A [PictureInfoDecoder] for [Uint8List]s that will clip to the viewBox.
   static final PictureInfoDecoder<Uint8List> svgByteDecoder =
-      (Uint8List bytes, ColorFilter colorFilter, String key) =>
+      (Uint8List bytes, ColorFilter? colorFilter, String key) =>
           svg.svgPictureDecoder(bytes, false, colorFilter, key);
 
   /// A [PictureInfoDecoder] for strings that will clip to the viewBox.
   static final PictureInfoDecoder<String> svgStringDecoder =
-      (String data, ColorFilter colorFilter, String key) =>
+      (String data, ColorFilter? colorFilter, String key) =>
           svg.svgPictureStringDecoder(data, false, colorFilter, key);
 
   /// A [PictureInfoDecoder] for [Uint8List]s that will not clip to the viewBox.
   static final PictureInfoDecoder<Uint8List> svgByteDecoderOutsideViewBox =
-      (Uint8List bytes, ColorFilter colorFilter, String key) =>
+      (Uint8List bytes, ColorFilter? colorFilter, String key) =>
           svg.svgPictureDecoder(bytes, true, colorFilter, key);
 
   /// A [PictureInfoDecoder] for [String]s that will not clip to the viewBox.
   static final PictureInfoDecoder<String> svgStringDecoderOutsideViewBox =
-      (String data, ColorFilter colorFilter, String key) =>
+      (String data, ColorFilter? colorFilter, String key) =>
           svg.svgPictureStringDecoder(data, true, colorFilter, key);
 
   /// If specified, the width to use for the SVG.  If unspecified, the SVG
   /// will take the width of its parent.
-  final double width;
+  final double? width;
 
   /// If specified, the height to use for the SVG.  If unspecified, the SVG
   /// will take the height of its parent.
-  final double height;
+  final double? height;
 
   /// How to inscribe the picture into the space allocated during layout.
   /// The default is [BoxFit.contain].
@@ -574,13 +613,13 @@ class SvgPicture extends StatefulWidget {
   ///    specify an [AlignmentGeometry].
   ///  * [AlignmentDirectional], like [Alignment] for specifying alignments
   ///    relative to text direction.
-  final Alignment alignment;
+  final AlignmentGeometry alignment;
 
   /// The [PictureProvider] used to resolve the SVG.
   final PictureProvider pictureProvider;
 
   /// The placeholder to use while fetching, decoding, and parsing the SVG data.
-  final WidgetBuilder placeholderBuilder;
+  final WidgetBuilder? placeholderBuilder;
 
   /// If true, will horizontally flip the picture in [TextDirection.rtl] contexts.
   final bool matchTextDirection;
@@ -593,7 +632,7 @@ class SvgPicture extends StatefulWidget {
   ///
   /// The value indicates the purpose of the picture, and will be
   /// read out by screen readers.
-  final String semanticsLabel;
+  final String? semanticsLabel;
 
   /// Whether to exclude this picture from semantics.
   ///
@@ -601,13 +640,36 @@ class SvgPicture extends StatefulWidget {
   /// application.
   final bool excludeFromSemantics;
 
+  /// The content will be clipped (or not) according to this option.
+  ///
+  /// See the enum [Clip] for details of all possible options and their common
+  /// use cases.
+  ///
+  /// Defaults to [Clip.hardEdge], and must not be null.
+  final Clip clipBehavior;
+
+  /// The color filter, if any, to apply to this widget.
+  final ColorFilter? colorFilter;
+
+  /// Whether to cache the picture with the [colorFilter] applied or not.
+  ///
+  /// This value should be set to true if the same SVG will be rendered with
+  /// multiple colors, but false if it will always (or almost always) be
+  /// rendered with the same [colorFilter].
+  ///
+  /// If [Svg.cacheColorFilterOverride] is not null, it will override this value
+  /// for all widgets, regardless of what is specified for an individual widget.
+  ///
+  /// This defaults to false and must not be null.
+  final bool cacheColorFilter;
+
   @override
   State<SvgPicture> createState() => _SvgPictureState();
 }
 
 class _SvgPictureState extends State<SvgPicture> {
-  PictureInfo _picture;
-  PictureStream _pictureStream;
+  PictureInfo? _picture;
+  PictureStream? _pictureStream;
   bool _isListeningToStream = false;
 
   @override
@@ -639,11 +701,11 @@ class _SvgPictureState extends State<SvgPicture> {
   void _resolveImage() {
     final PictureStream newStream = widget.pictureProvider
         .resolve(createLocalPictureConfiguration(context));
-    assert(newStream != null);
+    assert(newStream != null); // ignore: unnecessary_null_comparison
     _updateSourceStream(newStream);
   }
 
-  void _handleImageChanged(PictureInfo imageInfo, bool synchronousCall) {
+  void _handleImageChanged(PictureInfo? imageInfo, bool synchronousCall) {
     setState(() {
       _picture = imageInfo;
     });
@@ -653,16 +715,16 @@ class _SvgPictureState extends State<SvgPicture> {
   // registration from the old stream to the new stream (if a listener was
   // registered).
   void _updateSourceStream(PictureStream newStream) {
-    if (_pictureStream?.key == newStream?.key) {
+    if (_pictureStream?.key == newStream.key) {
       return;
     }
 
     if (_isListeningToStream)
-      _pictureStream.removeListener(_handleImageChanged);
+      _pictureStream!.removeListener(_handleImageChanged);
 
     _pictureStream = newStream;
     if (_isListeningToStream) {
-      _pictureStream.addListener(_handleImageChanged);
+      _pictureStream!.addListener(_handleImageChanged);
     }
   }
 
@@ -670,7 +732,7 @@ class _SvgPictureState extends State<SvgPicture> {
     if (_isListeningToStream) {
       return;
     }
-    _pictureStream.addListener(_handleImageChanged);
+    _pictureStream!.addListener(_handleImageChanged);
     _isListeningToStream = true;
   }
 
@@ -678,7 +740,7 @@ class _SvgPictureState extends State<SvgPicture> {
     if (!_isListeningToStream) {
       return;
     }
-    _pictureStream.removeListener(_handleImageChanged);
+    _pictureStream!.removeListener(_handleImageChanged);
     _isListeningToStream = false;
   }
 
@@ -703,11 +765,24 @@ class _SvgPictureState extends State<SvgPicture> {
       );
     }
 
-    if (_picture != null) {
-      final Rect viewport = Offset.zero & _picture.viewport.size;
+    Widget _maybeWrapColorFilter(Widget child) {
+      // The color filter is already applied by the provider, or there is no
+      // color filter to apply at all.
+      if (widget.pictureProvider.colorFilter != null ||
+          widget.colorFilter == null) {
+        return child;
+      }
+      return UnboundedColorFiltered(
+        colorFilter: widget.colorFilter,
+        child: child,
+      );
+    }
 
-      double width = widget.width;
-      double height = widget.height;
+    if (_picture != null) {
+      final Rect viewport = Offset.zero & _picture!.viewport.size;
+
+      double? width = widget.width;
+      double? height = widget.height;
       if (width == null && height == null) {
         width = viewport.width;
         height = viewport.height;
@@ -718,18 +793,21 @@ class _SvgPictureState extends State<SvgPicture> {
       }
 
       return _maybeWrapWithSemantics(
-        SizedBox(
-          width: width,
-          height: height,
-          child: FittedBox(
-            fit: widget.fit,
-            alignment: widget.alignment,
-            child: SizedBox.fromSize(
-              size: viewport.size,
-              child: RawPicture(
-                _picture,
-                matchTextDirection: widget.matchTextDirection,
-                allowDrawingOutsideViewBox: widget.allowDrawingOutsideViewBox,
+        _maybeWrapColorFilter(
+          SizedBox(
+            width: width,
+            height: height,
+            child: FittedBox(
+              fit: widget.fit,
+              alignment: widget.alignment,
+              clipBehavior: widget.clipBehavior,
+              child: SizedBox.fromSize(
+                size: viewport.size,
+                child: RawPicture(
+                  _picture,
+                  matchTextDirection: widget.matchTextDirection,
+                  allowDrawingOutsideViewBox: widget.allowDrawingOutsideViewBox,
+                ),
               ),
             ),
           ),
@@ -740,12 +818,12 @@ class _SvgPictureState extends State<SvgPicture> {
     return _maybeWrapWithSemantics(
       widget.placeholderBuilder == null
           ? _getDefaultPlaceholder(context, widget.width, widget.height)
-          : widget.placeholderBuilder(context),
+          : widget.placeholderBuilder!(context),
     );
   }
 
   Widget _getDefaultPlaceholder(
-      BuildContext context, double width, double height) {
+      BuildContext context, double? width, double? height) {
     if (width != null || height != null) {
       return SizedBox(width: width, height: height);
     }